refactor(attention): name the released persona and test fixtures by convention
